+++ CHANGELOG.md
@@ -12,6 +12,7 @@ but this project DOES NOT adhere to [Semantic Versioning](http://semver.org/).
 * Remove now redundant calls to ConfigureAwait in async methods
 * Replace Newtonsoft.Json with System.Text.Json
 * `AsyncTasks.Hangfire` does not prepare schema by default now
+* `BackgroundProcessingApp` is now `HangfireTasksModule`
 
 ## 3.5
 * Rewrite of EFMigrator